create directors, backroomers
implement authorization checking

+++ src/main/java/ch/uzh/ifi/hase/soprafs26/controller/PlayerController.java
@@ -9,6 +9,7 @@ import ch.uzh.ifi.hase.soprafs26.rest.playerdto.RolePutDTO;
 import ch.uzh.ifi.hase.soprafs26.service.PlayerService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -20,9 +21,14 @@ public class PlayerController {
     public PlayerController(PlayerService playerService) {
         this.playerService = playerService;
     }
-    @PutMapping("/player/{playerid}")
-    public void assignUsertoPlayer(@PathVariable Long playerId, @RequestBody PlayerPutDTO playerPutDTO,  @RequestHeader("Authorization") String token) {
-        playerService.updatePlayerAssociation(playerId,playerPutDTO,token);
+    @PutMapping("/player/{playerId}")
+    public void assignUserToPlayer(@PathVariable Long playerId, @RequestBody PlayerPutDTO playerPutDTO,  @RequestHeader("Authorization") String token) {
+        String[] tokens = splitToken(token);
+        if(tokens[0].equals("Bearer")){
+            playerService.updatePlayerAssociation(playerId,playerPutDTO,tokens[1]);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PutMapping("/characters/{characterId}")
@@ -30,35 +36,48 @@ public class PlayerController {
     @ResponseBody
 
     public void updateRole(@RequestBody RolePutDTO rolePutDTO, @RequestHeader("Authorization") String token, @PathVariable Long characterId) {
-        playerService.updateRole(token, rolePutDTO,characterId);
+        String[] tokens = splitToken(token);
+        if(tokens[0].equals("Director")){
+            playerService.updateRole(tokens[1], rolePutDTO,characterId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @GetMapping("/characters/{characterId}")
     @ResponseStatus
     @ResponseBody
     public RoleGetDTO getRole(@RequestHeader("Authorization") String token, @PathVariable Long characterId) {
-        Role found = playerService.getRole(token,characterId);
-        return PlayerDTOMapper.INSTANCE.convertEntitytoRoleGetDTO(found);
+        String[] tokens = splitToken(token);
+        return PlayerDTOMapper.INSTANCE.convertEntitytoRoleGetDTO(playerService.getRole(tokens[1],characterId));
     }
 
     @PostMapping("/characters")
-    public RoleGetDTO createRole(@RequestBody RolePostDTO rolePostDTO, @RequestHeader("Authorization") String token){
-        Role created =playerService.createRole(token, rolePostDTO);
-        return PlayerDTOMapper.INSTANCE.convertEntitytoRoleGetDTO(created);
+    public RoleGetDTO createRole(@RequestBody RolePostDTO rolePostDTO, @RequestHeader("Authorization") String authToken){
+        String[] tokens = splitToken(authToken);
+        if(tokens[0].equals("Director")){
+            return PlayerDTOMapper.INSTANCE.convertEntitytoRoleGetDTO(playerService.createRole(authToken, rolePostDTO));
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/characters/{characterId}")
     public void deleteRole(@RequestHeader("Authorization")String token, @PathVariable Long characterId){
-        playerService.deleteRole(token,characterId);
-
+        String[] tokens = splitToken(token);
+        if (tokens[0].equals("Director")){
+            playerService.deleteRole(token,characterId);
+        } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
     }
 
-    @GetMapping("/character/{characterId}/interlocutors")
-
-    private ArrayList<String> stripToken(String token){
-        
+    protected String[] splitToken(String token){
+        if(token == null || token.isEmpty() || token.isBlank()){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Token is empty");
+        }
+        String[] thingy = token.split(" ");
+        if(thingy[0].equals("Role") || thingy[0].equals("Backroomer") || thingy[0].equals("Director")){
+            return thingy;
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Token is invalid");
     }
-
-
-
 }
